+++ lib/attributor/types/object.rb
@@ -14,19 +14,12 @@ module Attributor
       'An Object'
     end
 
-    # Not really used (we override as_json_schema to represent this as an Any Type),
-    # but if it _were_ used, this would be accurate.
-    def self.json_schema_type
-      :object
-    end
-
     # Represents Object as an OpenAPI Any Type.
+    # Which means there is no type key for an Object (i.e., 'any'), so we'll report it as nil
     #
     # @see https://swagger.io/docs/specification/data-models/data-types/#any
-    def self.as_json_schema(**kwargs)
-      schema = super(**kwargs)
-      schema.delete(:type)
-      schema
+    def self.json_schema_type
+      nil
     end
   end
 end